% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelowanie.R
\name{autokorelacja}
\alias{autokorelacja}
\title{Testy autokorelacji}
\usage{
autokorelacja(
  model,
  testy = c("Durbin-Watson", "Breusch-Godfrey"),
  rząd = 1,
  maksymalny = 1
)
}
\arguments{
\item{model}{Model liniowy, na którym mają być wykonane testy.}

\item{testy}{Wektor tekstowy z nazwami testów do wykonania. Domyślnie są to wszystkie dostępne testy.}

\item{rząd}{Określa rząd autokorelacji do testowania w teście Breusch-Godfrey. Domyślnie jest to 1.}

\item{maksymalny}{Określa maksymalny rząd autokorelacji do testowania w teście Breusch-Godfrey. Domyślnie jest to 1.}
}
\value{
Ramka danych z wynikami testów. Kolumny to: "statystyka", "p". Wiersze to nazwy wykonanych testów.
}
\description{
Funkcja wykonuje testy autokorelacji na podanym modelu. Dostępne testy to: "Durbin-Watson" i "Breusch-Godfrey".
}
\examples{
# Utwórz model liniowy na podstawie zestawu danych mtcars
model <- lm(mpg ~ ., data = mtcars)
# Wykonaj testy autokorelacji
wyniki_autokorelacji <- autokorelacja(model)
}
