% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{chomikuj}
\alias{chomikuj}
\title{Tworzy ramkę danych}
\usage{
chomikuj(..., nazwy = NULL)
}
\arguments{
\item{...}{Argumenty, które mają zostać umieszczone w ramce danych.}
}
\value{
Ramka danych zawierająca podane argumenty.
}
\description{
Funkcja tworzy ramkę danych z podanych argumentów, nie sprawdzając nazw.
}
\examples{
# Utwórz ramkę danych z wektorami
df <- zramkuj(a = 1:3, b = c("A", "B", "C"))
}
