% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelowanie.R
\name{inflacja}
\alias{inflacja}
\title{Diagnostyka inflacji wariancji}
\usage{
inflacja(dane)
}
\arguments{
\item{dane}{Model liniowy, dla którego mają być obliczone współczynniki VIF.}
}
\value{
Ramka danych z wartościami VIF dla każdej zmiennej w modelu.
}
\description{
Funkcja oblicza współczynniki inflacji wariancji (VIF) dla modelu.
}
