% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{które}
\alias{które}
\title{Znajduje indeksy spełniające warunek}
\usage{
które(x, arr.ind = FALSE, useNames = TRUE)
}
\arguments{
\item{...}{Warunek do sprawdzenia.}
}
\value{
Wektor z indeksami spełniającymi warunek.
}
\description{
Funkcja zwraca indeksy elementów wektora spełniających podany warunek.
}
\examples{
# Utwórz wektor
v <- c(1, 2, 3, 4, 5)
# Znajdź indeksy elementów większych od 3
które(v > 3)
}
