% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{nazwy<-}
\alias{nazwy<-}
\title{Ustawia nazwy wierszy lub kolumn}
\usage{
nazwy(x, poziomo = TRUE) <- value
}
\arguments{
\item{x}{Ramka danych, której nazwy mają być ustawione.}

\item{value}{Wektor z nowymi nazwami.}

\item{poziomo}{Czy mają być ustawione nazwy kolumn. Domyślnie FALSE.}
}
\value{
Ramka danych z ustawionymi nazwami.
}
\description{
Funkcja ustawia nazwy wierszy lub kolumn ramki danych.
}
\examples{
# Utwórz ramkę danych
df <- data.frame(a = 1:3, b = c("A", "B", "C"))
# Ustaw nazwy kolumn
nazwy(df, poziomo = TRUE) <- c("nowa_a", "nowa_b")
}
