% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{nazwy}
\alias{nazwy}
\title{Pobiera nazwy wierszy lub kolumn}
\usage{
nazwy(..., poziomo = TRUE)
}
\arguments{
\item{poziomo}{Czy mają być zwrócone nazwy kolumn. Domyślnie FALSE.}
}
\value{
Wektor z nazwami wierszy lub kolumn.
}
\description{
Funkcja zwraca nazwy wierszy lub kolumn ramki danych.
}
\examples{
# Utwórz ramkę danych
df <- data.frame(a = 1:3, b = c("A", "B", "C"))
# Pobierz nazwy kolumn
nazwy(df, poziomo = TRUE)
}
