% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelowanie.R
\name{normalność}
\alias{normalność}
\title{Testy normalności}
\usage{
normalność(
  próbka,
  testy = c("Shapiro-Wilk", "Anderson-Darling", "Lilliefors", "Jarque-Bera",
    "D'Agostino", "Craméra-von Mises", "Kolmogorov-Smirnov")
)
}
\arguments{
\item{próbka}{Wektor danych, na których mają być wykonane testy.}

\item{testy}{Wektor nazw testów do wykonania. Domyślnie wykonuje wszystkie dostępne testy.}
}
\value{
Ramka danych z wartościami statystyk testowych i wartościami p dla każdego wykonanego testu.
}
\description{
Funkcja wykonuje wybrane testy normalności na podanej próbce.
}
