% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelowanie.R
\name{ocena}
\alias{ocena}
\title{Ocena modelu liniowego}
\usage{
ocena(model)
}
\arguments{
\item{model}{Model liniowy do oceny.}
}
\value{
Ramka danych z wynikami oceny. Kolumny to: "R^2", "R^2 skoryg.", "p \link{F}", "AIC", "BIC".
}
\description{
Funkcja ocenia model liniowy, obliczając różne statystyki, takie jak R^2, skorygowane R^2, p-wartość dla statystyki F, AIC i BIC.
}
\examples{
# Utwórz model liniowy na podstawie zestawu danych mtcars
model <- zamodeluj(mpg ~ ., data = mtcars)
# Oceń model
wyniki <- ocena(model)
}
