% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelowanie.R
\name{parametry}
\alias{parametry}
\title{Pobierz parametry modelu liniowego}
\usage{
parametry(model)
}
\arguments{
\item{model}{Model liniowy, z którego mają być pobrane parametry.}
}
\value{
Ramka danych z parametrami modelu. Dla każdego parametru tworzone są dwie kolumny: jedna dla wartości parametru i druga dla wartości p.
}
\description{
Funkcja zwraca parametry modelu liniowego jako ramkę danych. Dla każdego parametru tworzone są dwie kolumny: jedna dla wartości parametru i druga dla wartości p.
}
\examples{
# Utwórz model liniowy na podstawie zestawu danych mtcars
model <- zamodeluj(mpg ~ ., data = mtcars)
# Pobierz parametry modelu
parametry_modelu <- parametry(model)
}
