% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{połącz}
\alias{połącz}
\title{Dodaje ramki danych}
\usage{
połącz(..., poziomo = FALSE)
}
\arguments{
\item{...}{Ramki danych do dodania.}

\item{poziomo}{Czy ramki danych mają być dodane w poziomie. Domyślnie FALSE.}
}
\value{
Ramka danych zawierająca dodane ramki danych.
}
\description{
Funkcja dodaje ramki danych w pionie lub w poziomie.
}
\examples{
# Utwórz dwie ramki danych
df1 <- data.frame(a = 1:3, b = c("A", "B", "C"))
df2 <- data.frame(a = 4:6, b = c("D", "E", "F"))
# Dodaj ramki danych w pionie
df <- dodaj(df1, df2)
}
