% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelowanie.R
\name{resztkuj}
\alias{resztkuj}
\title{Pobierz rezydua modelu liniowego}
\usage{
resztkuj(object, ...)
}
\arguments{
\item{model}{Model liniowy, z którego mają być pobrane rezydua.}
}
\value{
Wektor rezyduów z modelu.
}
\description{
Funkcja zwraca rezydua z modelu liniowego.
}
\examples{
# Utwórz model liniowy na podstawie zestawu danych mtcars
model <- zamodeluj(mpg ~ ., data = mtcars)
# Pobierz rezydua modelu
rezydua_modelu <- rezydua(model)
}
