% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{rysuj_próbkę}
\alias{rysuj_próbkę}
\title{Podgląd danych}
\usage{
rysuj_próbkę(
  próbka,
  słupki = 10,
  y = "Wartość",
  gęstość = FALSE,
  tytuł = "Wgląd w próbkę"
)
}
\arguments{
\item{próbka}{Wektor danych, dla których mają być stworzone wykresy.}

\item{słupki}{Liczba słupków w histogramie. Domyślnie 10.}

\item{y}{Etykieta dla osi y. Domyślnie "Wartość".}

\item{gęstość}{Czy dodać krzywą gęstości do histogramu. Domyślnie FALSE.}
}
\value{
Wykres ggplot z histogramem i boxplotem.
}
\description{
Funkcja tworzy histogram i boxplot dla podanej próbki.
}
