% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelowanie.R
\name{siatka}
\alias{siatka}
\title{Siatka wykresów}
\usage{
siatka(..., kolumny = NULL, rzędy = NULL)
}
\arguments{
\item{...}{Wykresy do umieszczenia w siatce.}

\item{kolumny}{Liczba wierszy w siatce. Domyślnie NULL.}

\item{rzędy}{Liczba kolumn w siatce. Domyślnie NULL.}
}
\value{
Siatka wykresów.
}
\description{
Funkcja tworzy siatkę wykresów na podstawie podanych wykresów.
}
