% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelowanie.R
\name{statystyki}
\alias{statystyki}
\title{Statystyki próbki}
\usage{
statystyki(próbka)
}
\arguments{
\item{próbka}{Wektor danych, dla których mają być obliczone statystyki.}
}
\value{
Ramka danych z obliczonymi statystykami.
}
\description{
Funkcja oblicza podstawowe statystyki dla podanej próbki.
}
