% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testuj.R
\name{testuj_homoskedastyczność}
\alias{testuj_homoskedastyczność}
\title{Testy homoskedastyczności}
\usage{
testuj_homoskedastyczność(model, testy = c("Breusch-Pagan", "Goldfeld-Quandt"))
}
\arguments{
\item{model}{Model liniowy, na którym mają być wykonane testy.}

\item{testy}{Wektor tekstowy z nazwami testów do wykonania. Domyślnie są to wszystkie dostępne testy.}
}
\value{
Ramka danych z wynikami testów. Kolumny to: "statystyka", "p". Wiersze to nazwy wykonanych testów.
}
\description{
Funkcja wykonuje testy homoskedastyczności na podanym modelu. Dostępne testy to: "Breusch-Pagan", "White" i "Goldfeld-Quandt".
}
