% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testuj.R
\name{testuj_liniowość}
\alias{testuj_liniowość}
\title{Testy założenia liniowości}
\usage{
testuj_liniowość(model, testy = c("Rainbow", "RESET", "Harvey-Collier"))
}
\arguments{
\item{model}{Model liniowy, na którym mają być wykonane testy.}

\item{testy}{Wektor tekstowy z nazwami testów do wykonania. Domyślnie są to wszystkie dostępne testy.}
}
\value{
Ramka danych z wynikami testów. Kolumny to: "statystyka", "st. swobody [1]", "st. swobody [2]" nie dotyczy Harv.-Coll., "p". Wiersze to nazwy wykonanych testów.
}
\description{
Funkcja wykonuje testy liniowości na podanym modelu. Dostępne testy to: "Rainbow", "RESET" i "Harvey-Collier", patrz \code{lmtest}.
}
