% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{wczytaj}
\alias{wczytaj}
\title{Wczytuje dane z pliku CSV}
\usage{
wczytaj(
  file,
  header = TRUE,
  sep = ",",
  quote = "\\"",
  dec = ".",
  fill = TRUE,
  comment.char = "",
  ...
)
}
\arguments{
\item{file}{Nazwa pliku CSV do wczytania.}
}
\value{
Ramka danych z wczytanymi danymi.
}
\description{
Funkcja wczytuje dane z pliku CSV i zwraca ramkę danych.
}
\examples{
# Wczytaj dane z pliku "dane.csv"
dane <- wczytaj("dane.csv")
}
