% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{wektoryzuj}
\alias{wektoryzuj}
\title{Tworzy wektor}
\usage{
wektoryzuj()
}
\arguments{
\item{...}{Argumenty, które mają zostać umieszczone w wektorze.}
}
\value{
Wektor zawierający podane argumenty.
}
\description{
Funkcja tworzy wektor z podanych argumentów.
}
\examples{
# Utwórz wektor z liczbami od 1 do 3
wektor <- wektoryzuj(1, 2, 3)
}
