% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelowanie.R
\name{zamodeluj}
\alias{zamodeluj}
\title{Wykonaj model liniowy używając \code{lm}}
\usage{
zamodeluj(...)
}
\arguments{
\item{...}{Argumenty przekazywane do funkcji \code{lm}.}
}
\value{
Model liniowy stworzony za pomocą funkcji \code{lm}.
}
\description{
Funkcja tworzy model liniowy na podstawie podanych argumentów, używając funkcji \code{lm} z R.
}
\examples{
# Utwórz model liniowy na podstawie zestawu danych mtcars
model <- zamodeluj(mpg ~ ., data = mtcars)
}
