% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\name{zlistuj}
\alias{zlistuj}
\title{Tworzy listę}
\usage{
zlistuj()
}
\arguments{
\item{...}{Argumenty, które mają zostać umieszczone w liście.}
}
\value{
Lista zawierająca podane argumenty.
}
\description{
Funkcja tworzy listę z podanych argumentów.
}
\examples{
# Utwórz listę z liczbami od 1 do 3
lista <- zlistuj(1, 2, 3)
}
